microservices/queue-service/requirements.txt
Django>=4.0
celery[redis]
djangorestframework
redis
